# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Qberry Laravel API documentation",
        "contact": {
            "email": "bc.serj@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1/",
            "description": "Qberry Laravel API server"
        }
    ],
    "paths": {
        "/locations": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Get all locations",
                "operationId": "locationsAll",
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Store new location",
                "description": "Return created location data",
                "operationId": "locationAdd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreLocationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/locations/{id}": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Get location information",
                "description": "Returns location data",
                "operationId": "locationGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Location id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Location"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/timezones": {
            "get": {
                "tags": [
                    "Timezone"
                ],
                "summary": "Get all timezones",
                "operationId": "timezonesAll",
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/timezones/{id}": {
            "get": {
                "tags": [
                    "Timezone"
                ],
                "summary": "Get timezone data",
                "operationId": "timezoneGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Timezone id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Location": {
                "title": "Location",
                "description": "Location model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Location title",
                        "type": "string",
                        "example": "Kiyv"
                    },
                    "timezone": {
                        "$ref": "#/components/schemas/Timezone"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Location"
                }
            },
            "Timezone": {
                "title": "Timezone",
                "description": "Timezone model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "title",
                        "description": "Time zone",
                        "type": "string",
                        "example": "Europe/Kiev"
                    },
                    "offset": {
                        "title": "offset",
                        "description": "Time zone offset",
                        "example": "+03:00"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Timezone"
                }
            },
            "LocationResource": {
                "title": "LocationResource",
                "description": "Location resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Location"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LocationResource"
                }
            },
            "LocationsCollection": {
                "title": "LocationsCollection",
                "description": "Locations collection",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Location"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "LocationResource"
                }
            },
            "StoreLocationRequest": {
                "title": "Store Location request",
                "description": "Store Location request body data",
                "required": [
                    "title",
                    "timezone_id"
                ],
                "properties": {
                    "title": {
                        "title": "title",
                        "description": "Name of the new location",
                        "example": "Kiyv"
                    },
                    "timezone_id": {
                        "title": "timezone_id",
                        "description": "Timezone`s id of the new location",
                        "example": "337"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-API-TOKEN": {
                "type": "apiKey",
                "name": "X-API-TOKEN",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": ""
        },
        {
            "name": "Timezone",
            "description": ""
        },
        {
            "name": "Location",
            "description": ""
        }
    ]
}